+++ apps/stripe/src/app/api/saleor/payment-gateway-initialize-session/use-case.ts
@@ -9,7 +9,7 @@ import {
   BrokenAppResponse,
 } from "@/modules/saleor/saleor-webhook-responses";
 
-import { responseData } from "./response-data";
+import { responseData } from "./response-params";
 import {
   PaymentGatewayInitializeSessionUseCaseResponses,
   PaymentGatewayInitializeSessionUseCaseResponsesType,
